nfc: move poller defenition files to protocols
