+++ lib/nfc/nfc_poller_manager.c
@@ -1,5 +1,6 @@
 #include "nfc_poller_manager.h"
-#include "nfc_poller_defs.h"
+
+#include <nfc/protocols/nfc_poller_defs.h>
 
 #include <furi.h>
