+++ lib/nfc/protocols/iso14443_4a/iso14443_4a_poller_defs.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include <nfc/nfc_poller_base.h>
+#include <nfc/protocols/nfc_poller_base.h>
 
 extern const NfcPollerBase nfc_poller_iso14443_4a;
