+++ lib/nfc/protocols/mf_desfire/mf_desfire_poller_defs.h
@@ -1,5 +1,5 @@
 #pragma once
 
-#include <nfc/nfc_poller_base.h>
+#include <nfc/protocols/nfc_poller_base.h>
 
 extern const NfcPollerBase mf_desfire_poller;
